requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord==1.0.1
discord.py==1.6.0
idna==3.1
multidict==5.1.0
Pillow==8.1.0
typing-extensions==3.7.4.3
Wand==0.6.5
yarl==1.6.3
